@@ -8,7 +8,6 @@
   "dependencies": {
     "@mui/icons-material": "^5.11.0",
     "@mui/material": "^5.11.3",
-    "@mui/utils": "^5.11.2",
     "@matrix-widget-toolkit/api": "3.1.0",
     "@matrix-widget-toolkit/mui": "1.0.7",
     "@matrix-widget-toolkit/react": "1.0.4",
@@ -19,8 +18,8 @@
     "joi": "^17.7.0",
     "lodash": "^4.17.21",
     "matrix-widget-api": "^1.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^12.0.0",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.8.1",
@@ -37,6 +36,9 @@
     "test": "jest --watch",
     "depcheck": "depcheck --ignores=@types/jest,@types/node,i18next-parser --ignore-dirs=build",
     "translate": "i18next src/**/*.{ts,tsx}",
+    "pack": "echo \"Nothing to pack\"",
+    "prepack": "echo \"Nothing to prepack\"",
+    "postpack": "echo \"Nothing to postpack\"",
     "check-api-report": "echo \"Nothing to report\"",
     "generate-api-report": "echo \"Nothing to report\""
   },
@@ -56,13 +58,13 @@
     "@craco/craco": "^6.4.5",
     "@matrix-widget-toolkit/testing": "2.2.0",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^27.5.2",
     "@types/jest-axe": "^3.5.5",
     "@types/node": "^16.18.11",
-    "@types/react": "^17.0.45",
-    "@types/react-dom": "^17.0.16",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
     "i18next-parser": "^7.6.0",
     "jest-axe": "^7.0.0",
     "msw": "^1.0.0",
