@@ -16,8 +16,8 @@
     "joi": "^17.7.0",
     "lodash": "^4.17.21",
     "matrix-widget-api": "^1.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-i18next": "^12.0.0",
     "react-redux": "^8.0.5",
     "react-router-dom": "^6.8.1",
@@ -35,6 +35,9 @@
     "test": "jest --watch",
     "depcheck": "depcheck --ignores=@types/jest,@types/node,i18next-parser --ignore-dirs=build",
     "translate": "i18next src/**/*.{ts,tsx}",
+    "pack": "echo \"Nothing to pack\"",
+    "prepack": "echo \"Nothing to prepack\"",
+    "postpack": "echo \"Nothing to postpack\"",
     "check-api-report": "echo \"Nothing to report\"",
     "generate-api-report": "echo \"Nothing to report\""
   },
@@ -55,12 +58,12 @@
     "@matrix-widget-toolkit/testing": "2.2.0",
     "@semantic-ui-react/craco-less": "^3.0.0",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.11",
-    "@types/react": "^17.0.45",
-    "@types/react-dom": "^17.0.16",
+    "@types/react": "^18.0.28",
+    "@types/react-dom": "^18.0.11",
     "i18next-parser": "^7.6.0",
     "msw": "^1.0.0",
     "typescript": "^4.8.4"
