@@ -10,16 +10,15 @@
   "devDependencies": {
     "@craco/craco": "^6.4.5",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^27.5.2",
     "@types/jest-axe": "^3.5.5",
     "@types/node": "^16.18.11",
-    "@types/react": "^17.0.45",
+    "@types/react": "^18.0.28",
     "copyfiles": "^2.4.1",
     "i18next-parser": "^7.6.0",
     "jest-axe": "^7.0.0",
-    "react": "^17.0.2",
     "react-scripts": "5.0.1",
     "typescript": "^4.8.4"
   },
@@ -50,7 +49,7 @@
     "i18next-browser-languagedetector": "^7.0.1",
     "i18next-resources-to-backend": "^1.0.0",
     "matrix-widget-api": "^1.1.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-i18next": "^12.0.0",
     "react-use": "^17.4.0"
   },
