@@ -10,12 +10,11 @@
   "devDependencies": {
     "@craco/craco": "^6.4.5",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
-    "@testing-library/react-hooks": "^8.0.1",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.11",
-    "@types/react": "^17.0.45",
+    "@types/react": "^18.0.28",
     "react-scripts": "5.0.1",
     "typescript": "^4.8.4"
   },
@@ -34,7 +33,7 @@
   "dependencies": {
     "@matrix-widget-toolkit/api": "^3.0.0",
     "matrix-widget-api": "^1.1.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-error-boundary": "^3.1.4",
     "react-use": "^17.4.0"
   },
