@@ -12,12 +12,11 @@
     "@craco/craco": "^6.4.5",
     "@semantic-ui-react/craco-less": "^3.0.0",
     "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^12.1.5",
+    "@testing-library/react": "^14.0.0",
     "@testing-library/user-event": "^14.4.3",
     "@types/jest": "^27.5.2",
     "@types/node": "^16.18.11",
-    "@types/react": "^17.0.45",
-    "react": "^17.0.2",
+    "@types/react": "^18.0.28",
     "copyfiles": "^2.4.1",
     "i18next-parser": "^7.6.0",
     "react-scripts": "5.0.1",
@@ -44,7 +43,7 @@
     "i18next-browser-languagedetector": "^7.0.1",
     "i18next-resources-to-backend": "^1.0.0",
     "matrix-widget-api": "^1.1.1",
-    "react": "^17.0.2",
+    "react": "^18.2.0",
     "react-i18next": "^12.0.0",
     "react-use": "^17.4.0",
     "semantic-ui-less": "^2.5.0",
